Travis CI deployment.

+++ scripts/deploy.sh
@@ -1,9 +1,8 @@
 #!/bin/bash
-set -x
-#if [ $TRAVIS_BRANCH == 'master' ] ; then
-    # Initialize a new git repo in _site, and push it to our server.
+
+if [ $TRAVIS_BRANCH == 'master' ] ; then
     git remote add deploy "deploy@katy-ereira.co.uk:/var/repo/activity-monitor.git"
     git push --force deploy $TRAVIS_BRANCH
-#else
-#    echo "Not deploying, since this branch isn't master."
-#fi
+else
+    echo "Not deploying, since this branch isn't master."
+fi
